pyproject.toml
[project]
name = "db"
version = "0.1.0"
description = "DB"
authors = [
    {name = "c1lxz"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "sqlalchemy[asyncio] (>=2.0.44,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)"
]

[dependency-groups]
dev = [
    "fastapi[standard] (>=0.124.0,<0.125.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
